Change tracking: flag full object as changed
Flag an entire object as being changed with :all_changed!

The setting is preserved on the change-tracking hash.

If the entire object is thus marked, when converting to_simple
with :changed=>true option, all attributes will be output whether
or not they were actually changed.  This is accomplished by pruning
out the :changed option, so the altered options propagate down the
object graph.